refactor: isolate client-safe settings types
