+++ src/widgets/configSecrets.ts
@@ -3,6 +3,7 @@ import {
   type KokpitConfig,
   type Service,
 } from "@/config/schema";
+import type { ClientSafeSettings } from "./clientSafeSettings";
 import { getWidget } from "@/widgets";
 import { widgetCredentialScopesMatch } from "./credentialScope";
 import {
@@ -20,19 +21,6 @@ import {
 /** The sole browser-visible key for an opaque unknown-widget config. */
 export const UNKNOWN_WIDGET_CONFIG_REFERENCE_KEY =
   "__kokpit_widget_config_reference__";
-
-/**
- * The settings representation that may cross a server-to-client boundary.
- *
- * This is intentionally a distinct type even though its editable, non-secret
- * fields currently have the same shape as KokpitConfig. Callers should never
- * pass a raw KokpitConfig to a response or Client Component: constructing this
- * type through `toClientSafeSettings` is the security boundary that replaces
- * widget credentials with opaque, server-verifiable references.
- */
-export type ClientSafeSettings = Omit<KokpitConfig, "services"> & {
-  services: Service[];
-};
 
 export type WidgetSecretResolutionErrorCode =
   | "widget_secret_reference_invalid"
@@ -111,7 +99,7 @@ function getOpaqueConfigReference(
  * registry metadata, so integrations do not need key-name redaction lists.
  */
 export function toClientSafeSettings(config: KokpitConfig): ClientSafeSettings {
-  return {
+  const safeConfig = {
     ...config,
     services: config.services.map((service) => {
       const widget = service.widget;
@@ -158,10 +146,8 @@ export function toClientSafeSettings(config: KokpitConfig): ClientSafeSettings {
       };
     }),
   };
+  return safeConfig;
 }
-
-/** @deprecated Use the explicitly named server-to-client boundary helper. */
-export const redactWidgetSecrets = toClientSafeSettings;
 
 function resolveReference(
   referenceValue: unknown,
